create new post fragment and move a lot of the suffixes over to that
